chore: Clean code a bit more
