+++ examples/client_example.py
@@ -3,8 +3,8 @@ from generals.remote import autopilot
 import argparse
 
 parser = argparse.ArgumentParser()
-parser.add_argument("--user_id", type=str, default="user_id9l")
-parser.add_argument("--lobby_id", type=str, default="elj2")
+parser.add_argument("--user_id", type=str, default=...) # Register yourself at generalsio and use this id
+parser.add_argument("--lobby_id", type=str, default="psyo") # After you create a private lobby, copy last part of the url
 parser.add_argument("--agent_id", type=str, default="Expander") # agent_id should be "registered" in AgentFactory
 
 if __name__ == "__main__":
